[0001] chore(modeling): unify numeric types to double

+++ include/core/robot_state.hpp
@@ -8,8 +8,8 @@ namespace robot::core {
  * @brief Current state of the robot at a given time
  */
 struct RobotState {
-    std::vector<float> jointPositions;   /*!< Current joint positions */
-    std::vector<float> jointVelocities;  /*!< Current joint velocities */
+    std::vector<double> jointPositions;   /*!< Current joint positions */
+    std::vector<double> jointVelocities;  /*!< Current joint velocities */
 };
 
 } // namespace robot::core
